Removing older plugin
